Supporting __setitem__() and mixed-mode initialisation.
Arbitrary operations with __setitem__() can be performed.
Also slices can be used in keys.

The test suite has been expanded to take account of the new
mixed-mode initialisation scheme with list-like structures, and
to test __setitem__().

Mixed-mode initialisation made changes in core.py necessary,
but all changes related to __setitem__ support are capsuled
in the machinery beneath and are not visible from outside.