Support for printing of undarrays and many minor things.
There is now support for printing of arbitrarily shaped undarrays
via the numpy printing engine and the upy.printing module.  The
printout can be modified by arguments to

	ua.printable()

When __str__() is invoked, default parameters are substituted.
There is support for two formats, 'float' and 'exp', and for
enforcing signs of the value and the exponent.  The general
printing scheme has been changed to (1.23 +- 0.45) 10^-6 or
(0.123 +- 0.045) respectively.

TODO:  Support for modifying the number of significant digits of
the uncertainty.

In upy.core, the preliminary printing support has been deleted,
together with the removal of the upy.decimal module, which has been
replaced by upy.decimal2.

ufuncWraps from upy.operators support now __str__ and __repr__,
beforehand they "inherited" this features from the original ufuncs
via __getattr__.

The upy.test script has been adapted to the new output syntax and
to the new functionality.  All current tests pass ...

There is now support for the arctan2 function in upy.umath.