Committing work stashed in 2011
I found the master branch with a stash. The stash dated back to
somewhere in the range [7 Feb 2011, 2 May 2011].  The lower limit is
the last commit on branch master, the upper limit is the first commit
on another branch 'maintenance/0.4.11b'.  I remember that the commits
on the maintenance branch were all later than the stash (noticed using
'gitk').

I would like to add to the changes of this commit: It is far from
stationary.  However, I keep it for the records and due to the severe
break of continuity which led to the present state.

I already added mini changes to 'LICENSE' and 'setup.py'.  I also
rearranged the directory structure to feature the 'upy' package as a
directory.