Work on Dependency
The naming scheme of the *names* and *derivatives* attributes of
:class:`Dependency` has been reverted from ``_names`` &
``_derivatives`` to ``names`` and ``derivatives``.
:meth:`Dependency.add` works with the *other*\ 's names & derivatives
directly.  Using "underscore" attributes from outside the object is
avoided by reverting the "underscorification".

Several occurences of ``copy`` have been corrected to ``other``.

A test in :meth:`add` has been introduced to make sure that the added
Dependency has the same shape as the indexed portion of *self*.

Masking by :meth:`__and__` now masks not only the names but also the
derivatives array.  Indeed :meth:`add` expects that *derivatives* is
zero where *names* is zero.