Changes to :class:`Characterstic`
The attribute ``_key_testing_array`` has been renamed to
``_shape_effect_array``.  It can be used in ``.compress``,
``.flatten``, and all methods of this sort to "simulate" the operation
and to calculate the output shape.  This is necessary because
instantiation of :class:`Characteristic` requires the desired shape.

Commented out some logic acting on *key* in :meth:`add`.  Formerly the
key has been transformed to a tuple if it hasn't been a tuple.
However, the usage of the *key* (also in Dependency) does not require
such munching.  As the logic used before can have side-effects, it is
removed.  The *key* is used now verbatim.