Simplifying code pathes
The code to multiply a Characteristic with *something else* has been
simplified a lot, using the shape effect array and relaying
broadcasting to :meth:`Dependency.__mul__`.

The implementation of :meth:`Characteristic.__rmul__` is now borrowed
from ``__mul__``, assuming a commuting product operation.

In :meth:`Dependency.__mul__`, ``self.names`` is now augmented by
adding an array of zeros with *other*\ 's shape in order to enforce
broadcasting of the shapes of *other* and ``self.names``.

:meth:`Dependency.__rmul__` is now borrowed from ``__mul__`` as well.