undarray: Duplicating creation of .shape and .ndim
``self.shape`` and ``self.ndim`` are now created separately in each
initialisation branch.  This is done because in the branch using only
*object* ("mixed-mode initialisation") these attributes need to be
created prior to the final initialisation step (``self[()] =
object``), and hence in this branch the initialisation would occur
twice.  It is not strictly necessary to do this change, it is rather a
matter of taste:  I do not want to treat the different branches
differently.