Substantial work in core.py
Introducing :func:`asundarray`, which acts like ``numpy.asarray``:  In
case the ``undarray_like`` argument is already an ``undarray``, it is
returned without copying; otherwise a new undarray is created.

Renaming the first positional argument of the ``undarray`` constructor
from ``object`` to ``master``.  The name "object" is just too general.
Also, "master" reflects that the argument is the "ruling" argument,
while keeping up with the ambiguity that it might be either an
undarray or the argument specifying the nominal value.

Reformatting the docstring of the ``undarray`` constructor.

Renaming the ``_characteristic`` attribute of :class:`undarray` to
``characteristic`` (without the leading underscore).  The
Characteristic instance is accessed from outside the respective
``undarray`` instance during initialisation in the branch dealing with
the *derivatives* argument.

Reordering the statements in many initialisation branches of
``undarray``.

I finally replaced the former initialisation branch called
"mixed-mode" initialisation by just ``numpy.asarray()``\ 'ing the
*master* argument in that case.  The code in
:meth:`undarray.__setitem__` has not been revisited yet.  Reason for
this transition is given in the previous commit.

In general, I changed code which took over arguments without copying
them such that the codepathes *do* copy these arguments.  The
transition is not yet done in all places.