Work on core.py
Introducing ``real`` and ``imag`` properties of :class:`undarray`.

Reformatting large sections of the code base: removing whitespace
around '=' signs.  However, it has been removed automatically by a
substitution command and has been applied in many places erroneously.
Apparently I've been rather tired.

Documented the :meth:`__pow__` code with a proof of the formulas used.
Same in :meth:`__rdiv__`.

Replaced the code for :meth:`__abs__`:  Before, real-valued comparison
using ``< 0`` has been used.  Now, the quotient of the absolute value
and the nominal value is used as a factor.  It is necessary to use
such a factor because the Characteristic needs to be updated.  The new
code can calculate this "adaption" factor also for complex-valued
nominal value; the replaced code was able to do this only for a
real-valued nominal value.