Finishing the transition work on 'core.py'
A slight update to the docstring of :meth:`undarray.__init__`.

Reformatting quite a lot of places to conform to the pattern of the
following example::

    return undarray(
        nominal=(self.nominal + other),     # properly use ( )
        derivatives=[(self, 1.0)],          # use a final ','
    )

:meth:`undarray.__setitem__` now is implemented more cleanly.  The
return value is assembled from a prepared nominal value and a prepared
Characteristic; the code no longer messes around in the internals of a
``undarray`` copy.

:meth:`undarray.cumprod` has been removed.  The implementation was not
elegant enough.  It was not the Right Thing (TM).  I also believe that
people probably won't need the ``cumprod`` method.

Other numpy.ndarray methods which still featured use of the "object"
argument to the ``undarray`` constructor have been updated to use
*nominal* instead.