On the way to refactoring the printing mechanism
One slight simplification in the :class:`undarray` constructor.

A difficult question in terms of protocol is the printing of
uncertainties.  The displayed uncertainty is a multiple of the
standard deviation, but the factor is not known a priori.

I tend to make the code agnostic w.r.t. the number of stddevs to be
used for uncertainty printout.  Unfortunataly, this makes standard
``str()`` or ``print uarray`` statements impossible to implement.
Hence I am returning now a symbolic string in
:meth:`undarray.__str__`.  As before, ``__repr__`` is not implemented,
as the result would need to be usable to reconstruct the ``undarray``
in question.

The constructor uses ``stddev`` explicitly, thus this code is already
agnostic w.r.t. the number of stddevs making up the error.

The use of a context manager is not truly helpful.  If the manager
shall be used automatically, a locking mechanism would be required.
If the manager is to be used explicitly, the manager can be replaced
by a simple variable.

However, it might be useful to provide an object which mangles
``undarray`` instances to strings.  Such an object could hold all
parameters for printing.  This would implement the DRY principle
(don't repeat yourself).  Furthermore, the user would be able to
centralise his choice of parameters, making the code more easy to
maintain and more terse.