Worked on typesetting framework
Introducing 'typesetters/{analysis.py, typesetter.py}':  Analysis and
typesetting of decimal numbers.

Using a module-scope instance of :class:`NoRule` in module 'adjstr.py'
to be shared amongst all plain 'str' instances passed through
:func:`asadjstr`.

Drafted the design of classes in 'typesetters/scientfic.py'.  Typset
numbers are converted to :class:`AdjustableString` instances by
instances of :class:`ScientificRule`.  Instances of
:class:`ScientificElement` delay typesetting the respective element.
Converting the numbers to :class:`TypesetNumber` instances is
performed by the :class:`ScientificTypesetter`, using the
:class:`ScientificRule` instance.

The AdjustableString rules cannot be stored in the
'ScientificTypesetter' because they are applicable to only one call of
'ScientificTypsetter.typeset(uarray)'.  Because of this, they are
stored in a 'ScientificRule' instance.

Each 'ScientificElement' keeps a record of the 'ScientificRule' to
apply to all of the "brother" instances of 'ScientificElement', and
hands the rule over to 'ScientificTypsetter.typeset_element()', in
order to typeset the respective element.

Using this design, all actual formatting rule is performed by the
'ScientificTypesetter'.  The 'ScientificRule' is a compound analogon
to the basic Rules in 'adjstr.py'.  :class:'ScientificElement' is
responsible for delaying the string conversion only.