Work on typesetting numbers and elements.
:class:`DecimalTypsetter` has been renamed to
:class:`NumberTypesetter`:  It is insignifcant that it typesets in
decimal, but it is significant that it does so only for singular
numbers.

'ceil'-ing is now part of the NumberTypesetter's initialisation.  It
is no longer argument to :meth:`typeset()` of the 'NumberTypesetter'.
Reason: Any instance of NumberTypesetter is used always for the same
kind of typesetting.  Both 'typeset_positive_sign' as well as 'ceil'
should hence be initialisation arguments ('typeset_positive_sign'
already is such an initialisation argument).

Began work on :meth:`ScientificTypesetter.typeset_element`:
Documenting the way it is done + trying get into gears with the
algorithm.