Work on the typesetting framework
This commit serves as a backup before attemting to make the code work.
There are changes:

*   In the AdjustableString framework;
*   Bugfixes in ``class NumberTypesetter``;
*   Adjustments in ``class ScientificRule`` (this portion of the code
    also needs adaption to changes in adjstr.py).

Most notably, the class hierarchy for adjustable strings has been
changed s.t. ``AdjustableString`` no longer assumes the presence of a
Rule always.  By this design change, a class ``VerbatimString`` can be
introduced, replacing the ``NoRule`` rule.  The code is benefiting
from this by an increase of clarity and beauty.  There is no more need
for a clumsy global ``NoRule`` instance.

There are also a coule of debug messages which will be cleaned away.

A couple of tests for the typesetting machinery have been implemented;
maybe they need accomodation to the recent changes of the tested code.

This commit also includes the ``str`` subclassing part of the Printing
Study.