Adjustments due to 'upy' -> 'upy2'
Changed '/setup.py' to reflect the new version '2-0.5.0a' and to
install the appropriate packages.

Adjusted all import statements and uses of the imported packages.

No other non-negligible content change yet.

A posteriori comment to 670d855d:

    Import statements which ran formerly 'import upy.<module>' cannot
    easily be translated to use 'upyb', as 'import upyb.<module> as
    upy.<module>' is not a valid import statement.  This increases the
    burden created by providing upy-0.4 under the name 'upyb'.

This indicates that the present direction is something like the Right
Thing (TM).