Scientific Typesetting reached beta stage
I re-implemented the Rule code inside of 'typesetting.scientific'.  The
concept of an "adjustable string" vanished.  Instead, Rules can be
applied to input data, possibly updating their internal state.

When the uarray in question is passed twice through ``str()``, the
results of the second pass use the final state of the Rules involved.

The resulting code is more elegant than the previous version.  This is
mainly due to the introduction of :class:`TypesetNumberRule`,
encapsulating typesetting of a Typeset Floating-Point Number.

I am providing some preliminary test codes in 'test/2016-08-07_1047/'.
Although they aren't exhaustive, they demonstrate that the typesetting
code is, at least in principle, functional.

There are three options of whitespace around typeset scientific
elements:

1.  No additional whitespace;
2.  one space at the end;
3.  Spaces before and after.

In (1.), elements are separated by the space introduced by numpy alone.
This is too narrow, because there are already spaces ' ' in the
scientific representation.

(2.) is okay.

(3.) is worse than (2.), as it prints leading spaces also for scalar
uarrays, what indents the output by one space (NOK).

The space at EOL does not matter too much.  In arrays, it separates the
trailing ']' by one space, and in scalar uarrays it does not harm
either.