First draft of Context functionality
I've chosen German language.  Expressing myself in German is more easy
for me; and if I ever write a documentation I can use it as resource
for an English text.  I also believe the first mind-storming draft
text is not too important - I will hopefully stay the only one who
needs to read it.  When things settled down, an appropriate commit
message will be written.

The draft text succeeds in finding an outline of the functionality
intended; the design of the actual registry framework is pending.