:class:`Context` reached Beta state
Renaming a lot of attributes for the sake of clarity.

Introducing the Default Stack.  Before only one default item could be
stored.  :meth:`default` and :meth:`undefault` as well as
:meth:`current` have been adapted accordingly.

Rendering access to the Default Stack atomic in all places.

Providing doc strings throughout.

Raising LookupError when both the Thread Stack as well as the Default
Stack are empty.  Before, ``None`` was returned.