Providing the ``u`` calculus
Introducing :class:`upy2.core.U`` and :func:`upy2.core.u`.  Both define
an "uncertainty calculus" by enabling the user to write code like::

    10 +- u(2)

or, more precisely::

    with U(2):  # Interpret errors as given in 2 sigma convention
        a = 10 +- u(2)

However, a problem rooted deeply inside of upy2 appeared.  The
abovementioned call fails, because inside of the upy2 core somewhere a
scalar array turned into a "true" scalar.  The difference is that the
scalar array is an instance of :class:`undarray`, while the "true"
scalar is, in our case, an instance of :class:`numpy.float64`.  The
former supports item assignment like ``a[()] = 42``, the latter lacks
such support.

Because of this, the statement::

    [10] +- u([2])

executes properly, while ``10 +- u(2)`` does not.  As soon as the shape
isn't equal to ``()``, the problem naturally disappears (such ``ndarray``
instances never can reduce to numbers like ``numpy.float64`` instances).

This problem is deeply rooted and needs *very* careful investigation, to
not leave any of such problematic code pathes uncorrected.  However,
when the offending assignments are identified, they probably can be
fixed easily by wrapping a call of ``numpy.asarray()`` around the
problemantic expression.

The ``u`` calculus has been tested in a shallow manner; so far it seems
functional.  The test code is located in
[/tests/drafts/2016-10-21_1954/a02.py].