Boiling down the problem reported in 459023a
The culprit statement of 459023a is a sum of two scalar arrays in
[dependency.py:144].  Notice that [dependency.py:144] is not the
expression which fails; that is [dependency.py:158].
[dependency.py:144] is the place where ``self.derivatives`` reduces to a
"true" scalar.  The expressions boil down to what is given in
[test/drafts/2016-1021_1954/a03.py]: wrapping the sum in
``numpy.asarray()`` will solve the problem.

However:

1.  Similar statements might be used at other places too; and
2.  there might be similar expression types (products, for example) which
    trigger the same error.

I need to scan the whole code of upy2 for such situations.