Contexts: Documentation and simplification
Decided to keep the name "Context"; thoughts have been put into the
study [2016-09-05_1205] (in German).  The decision seems robust.

Added and improved docstrings in 'context.py'.

No longer caching the Context in :class:`context.Protocol`.  The reason
that I did before was faulty.  Subclasses of the Protocol key won't
"shadow" the original key, since the Protocol object won't be an
instance of them.