Context: Decision w.r.t. Python Context Managers
Protocol Implementations, when used as Python Context Managers, can
*not* be used like::

    with U(2) as u:
        [...]

Their only supported use case is::

    with U(2):
        ... use :func:`upy2.core.u` ...

The rationale is given in [Studies/2016-09-05_1205/a01:293 ff.] (in
German).

The former use case is, according to that Study, never needed, and can
always be safely substituted by the latter or by not using the Protocol
Implementation as a Context Manager at all; even more, the latter use
pattern is superior to the former in all considered cases.