Operators: Committing messy state for the records
The explanation of the problem which is attempted to be remedied by upy
Operators has been removed in its entirety.  I've not been happy with
the text, and prefer a complete rewrite when it is time to.

The code implementing upy Operators is in early beta phase.  Naming
conventions did not stabilise yet.

All in all, I find this heritage of early upy ages difficult to
maintain.