Operators: argument renames and using numpy.add
I renamed the arguments from a,b to x1,x2.

The latter is the naming convention followed by numpy, so I parallel
that.  It has also the advantage that *output parameters* can be
numbered as well: ``out1, out2`` - for instance in ``divmod``, if this
ever comes to use.

From this commit on, the :class:`Add` uufunc uses ``numpy.add`` to add
the two nominal values ``y1`` and ``y2`` instead of adding them by
``+``.  This has the advantage that values for ``x1`` and ``x2`` which
are neither an ndarray nor an undarray (for example, lists) can be
handled properly.

Using ``numpy.add`` comes at the expense that the addition cannot be
overridden by the operands anymore.

The main use case will be to hand over (nested) lists and the like.
These require either to pass them through ``numpy.asarray`` or using
``numpy.add``.  Both of them disallow overriding the addition operation.

Latest in the ``undarray`` constructor the result of addition will be
passed through ``numpy.asarray``.

From the Operator Study it seems that upy won't support non-ndarray
nominal values.