Study on the behaviour of the new 'absolute'
This is a post-natem commit.

There has been the attempt of proving the observation that ``sqrt(ua *
ua.conjugate())`` returns uncertain ndarrays with zero imaginary part.

The proof is not complete.

However, the observation that the failure of calculating the absolute
value of complex uncertain ndarrays with zero nominal value is now
better understood and it is likely that this behaviour is in fact the
Right Thing (TM).