Work on :meth:`__setitem__` of :class:`undarray`
No longer constructing an intermediate ``undarray`` without uncertainty
when the *value* given isn't already an ``undarray``.  Instead, the two
cases (``undarray`` and non-``undarray``) are handled *completely* in
the respective branch.

The resulting code is more readable and is expected to be more quick in
execution.  It is also a bit shorter.