Work on :meth:`U.provide`: Dependency construction
:meth:`U.provide` now constructs the Dependency encoding the requested
error *directly* instead of leaving this task to
:meth:`undarray.__init__`.

The corresponding branch in :meth:`undarray.__init__` will most probably
be pruned.  Constructing an ``undarray`` from nominal values and
independent standard deviations is considered too special to be put into
the general ``undarray`` constructor; hence it is left over to the
Uncertainty Provider :class:`U`, which is meant for this purpose.

The code resulting from this approach is more concise than the
implementation via :meth:`undarray.__init__`.