Introducing ``upy2.core.unegative``
Providing in :mod:`upy2.core`:

*   :class:`Unary`, derived from :class:`uufunc`;
*   :class:`Negative`, derived from :class:`Unary`;
*   :attr:`unegative`, an instance of :class:`Negative`.

and calling upon ``unegative`` in :meth:`undarray.__neg__`.

Introducing in :mod:`upy2.numpy_operators`:

*   :class:`UnaryOperator`, paralleling :class:`BinaryOperator`; and
*   instantiating :class:`UnaryOperator` in :func:`install_numeric_ops`
    for use with the previous 'negative' numpy numeric op and the
    ``unegative`` uufunc.

Testing has been performed in a *very* shallow manner by running
[tests/ad-hoc/2018-01-17_0926.py].  At least, that piece of code is
executing properly.

``unegative`` has been chosen for implemenation because it complements
the existing addition uufunc in::

    ua = ar +- u(err)

so it was needed in oder to be able to test ``uadd`` and ``u`` in the
way they are meant to be used.