Introducing subtraction uufunc
uufunc introduced in :mod:`core` and used in :mod:`numpy_operators`.

This is the first commit where the previous implementation of
``undarray`` operators like :meth:`undarray.__add__` is *removed in its
entirety*.  This heritage will be *removed* step by step beginning with
this commit.  It *won't* be kept as comments.  It will only be
accessible via git.