Work on elimination of :class:`Characteristic`
:mod:`upy2.characteristic` is no longer imported in :mod:`upy2.core`.

The following methods have been updated to use *self.dependencies*
instead of *self.characteristic*:

-   ``compress``
-   ``copy``
-   ``flatten``
-   ``repeat``
-   ``reshape``
-   ``transpose``

The new codes are a little longer than the original codes, while
incorporating the whole of :class:`Characteristic`.

This commit completes the elimination of :class:`Characteristic` in
:class:`undarray`, except the use in :meth:`undarray.__init__`.

:meth:`undarray.__init__` will undergo refactoring once the operators
are ready to go.  The invalid pathes will most likely be removed.