New approach to calculating the absolute value
Post-natem commit.

Added a new ad-hoc test for a novel notion to implement the calculation
of the absolute value by using ``sqrt(ua * ua.conj())``.  This approach,
borrowed from complex numbers, yields surprisingly beautiful results
when using it on undarrays.

The new approach is implemented for the purpose of testing in
[tests/ad-hoc/2018-05-09_1108.py].  It is tested there for various
single-dependency inputs, and the results are stunning:

*   Uncertainties with the same direction in the complex plane result in
    an appropriate uncertainty of the absolute value::

        ua = 1j +- u(0.2j)

        abs(ua) = 1 +- 0.2

*   Uncertainties which *do not alter* the magnitude of the uncertain
    value in first order::

        ua = 1j +- 0.2

    result in a precise absolute value::

        abs(ua) = 1

    where the error is zero.

*   Dependencies wich are "oblique" to the nominal value::

        ua = 1 + 1j +- u(0.2)   or   ua = 1 + 1j +- u(0.2j)

    result in a uncertainty of the absolute value which is smaller in
    magnitude than the error magnitude of the argument ``ua``::

        abs(ua) = 1.41 +- 0.15  (truncated to two decimals)

    where ``0.2 / sqrt(2) = 0.14142135...``.

The imaginary components of the so-produced absolute value candidates
vanish *in all of these tests*.

More tests are needed, and possibly a proof of this property is needed,
but already now the new approach seems more sensible than the one which
is implemented; and the new notion will result in much more clear and
elegant codes.