Further tests of the new approach to 'absolute'
The existing tests in [test/ad-hoc/2018-05-09_1108.py] already indicated
that the ``sqrt(ua * ua.conjugate())`` possibly makes sense, and further
tests harden this assumption: Also when making the operand depend on
*more than one* variable, with (reproducible) random complex values for
nominal value and derivative, the imaginary components of the absolute
result candidates *vanish always*.

Notice that the calculation of the absolute value of uncertain zeros
(e.g., ``0 +- 1`` or ``0 +- 1j``) *fails*.  The function succeeds with
plain numbers as operands.  More research is needed on this.  Notice
that the current implementation of :func:`absolute` uses ``<...> ** 0.5``
instead of ``sqrt()``.

Technical details:

1.  Using ``.conjugate()`` instead of ``.conj()``, as the former exists
    also for native Python numbers (complex numbers, but also ints and
    floats).

2.  Introducing :meth:`undarray.conjugate`; :meth:`undarray.conj` is a
    shorthand for this.