Work mostly on the calculation of absolute values of complex undarrays
It seems my study of the absolute value calculation of an undarray
``ua`` based on::

    sqrt(ua * ua.conj())

has come to some usable result.  It is described in German within
[studies/2017-11-03_1721/a17.txt].

The result is beautifully short, and is well explained in terms of
interpretation.  With these achievement, the old codes related to the
former :class:`Absolute` were pruned away.

During formulation of the study's results I converted
:meth:`undarray.depend` to :meth:`undarray.copy_dependencies`.  This led
to further simplifications.  The codes of :class:`Unary` and
:class:`Binary` were modified s.t. the new algorithm for calculating the
absolute value can be formulated.  The new interface is a generalisation
of the former one.  The new operators definitely need testing.

Some other changes to :file:`setup.py` are committed here as well:

1.  Replaced the commented-out url referring to sourceforge by the URL
    of the ``upy`` reposority.  Although this is not well-suited for new
    users, it is the best I can offer at the moment, given that the
    existing ``upy`` sites are out of date.
2.  Removed the "natural language" trove classifier.  As I see it, it
    just doesn't apply.