Combo update
Changed the name of the project in :file:`setup.py` from ``upy`` to
``upy2``.

Migrated the tests regarding :mod:`typesetting` in
:file:`/test/suite/test.py`.

Simplified the initialisers of :class:`undarray` and of
:class:`Dependency` significantly.  The *dtype* argument can now be used
to override the data type of the respective object.  Since the new codes
pass the tests now, the stale code pathes are to be removed and the
docstrings are to be updated.

Fixed a typo in :mod:`numpy_operators`.

In :meth:`ScientificTypesetter.typeset`, I am using now an iterator to
created the object-ndarray holding the :class:`ScientificElement`
instances.  Before, I ravelled the ndarrays involved and reshaped them
after having iterated them.  The latter works as well, also with non-C
strided arrays (:meth:`flatten` always traverses the operand C-order),
but the new approach is quite a bit more elegant, avoiding the
flattening step.  There is a test verifying that things are typeset
properly also with a non-C strided nominal-value ndarray.

All existing tests are passing.  There is still no black-box test
covering the actual uncertainty propagation.