Dependency: Removing support for :meth:`__rmul__`
The reliability of reverse multiplication would be flaky.  In some cases
it would work, but with ndarrays as other operand would would *not*
work.  Details of the results with :meth:`__rmul__` defined are given in
:file:`/test/suite/dependency.py`.

Tested a number of operations with a ``Dependency`` as second operand to
raise an appropriate Exception (``TypeError``).

With this commit :class:`Dependency` is now derived from ``object``.