Bugfix in :class:`core.{Unary, Binary}`
:class:`Unary` and :class:`Binary` now pass non-``undarray`` arguments
through :func:`numpy.asarray`.

This has been changed as :meth:`core.Divide._source1` calculates
``1.0/y2``, which fails when ``y2`` isn't an ``undarray`` (but,
possibly, a list).

The problem could also be handled locally in :meth:`Divide._source1`,
however, I deem it more clean to adopt the ``ndarray`` convention for
*all* uufuncs instead.

Introducing some more testing routines, which also triggered the bug.
These tests only verify that the operators work for list-type and
``ndarray``-type arguments, in the sense that they execute and return an
``undarray``.