Adapted the test suite s.t. it runs both with Python2 as well as with Python3
The actual ``upy2`` package didn't need any modification to be Python3
compatible.

Notice that in Python3 ``range`` is equivalent to the Python2
``xrange``.  Using the classical ``range`` in Python2 doesn't harm much.

Notice further that in Python3, there is no "div" (but only "truediv"
and "floordiv"), so the "div" tests only pertain to Python2.

All Tests are passing.