Augmented the test of :meth:`undarray.copy_dependencies`
The additional assertions test that the behaviour with a dependency
source featuring a too large shape is as intended.

When there are no Dependencies in the *soure*, *no* exception is raised
in that case.  I am refraining from implementing some "dry run" check,
because it might not cover all possible *key* arguments.

:meth:`undarray.copy_dependencies` is called currently:

*   In :meth:`undarray.__setitem__`, here the call will fail with too
    large *value* during the assignment to ``self.nominal``;

*   in :meth:`Unary.__call__` and meth:`Binary.__call__`, in this case
    the Dependency target has a shape large enough for both operands
    always.

The additional tests introduced here are thus merely a safety measure.
The abovementioned "dry run" checks can safely be left unimplemented.