Rewording of the *long description*
Returning to the wording of the sentence w.r.t. the mathematical
functions supported as it was in 6d838c.  Nowadays in fact "a range of
mathematical operations is supported".

Removing the notice about thread safety.  I am by intention *not*
protecting the non-atomic operations by locks, as locking is expensive.
I leave it to the user to render such operations transactional.

The contents of :file:`/README.rst` and :file:`/setup.py` are *in sync*.