Made the arccosh Study program :file:`/test/studies/arccosh.py` Python3-ready
Converted the ``print`` statements such that they can be run by Python 3
as calls of the ``print`` function while maintaining compatibility to
Python 2.

Slight improvement of the message reporting failure to calculate the
numerical differentiation (this message never applies within the tests).

Using ``range`` instead of ``xrange``.  The Python 3 ``range`` is
equivalent to the Python 2 ``xrange``, and the arguments are small
enough to allow using the "classic" ``range`` in Python 2.

Used the ``2to3`` helper script to provide the starting point of this
alteration.