Made the ad-hoc Tests ready for Python 3
Finally I decided to use ``2to3`` also on the ad-hoc Tests, even though
they should not be run by Python 2 after this step (sometimes ``print``
is called with more than one argument, which looks in Python 2 like
printing a tuple).

I verified that the output of the ad-hoc Tests didn't change with the
transition from Python 2 to Python 3 (Study [2020-09-04_0736]).