Introducing engineering typesetting
Providing :mod:`typesetting.engineering` with :class:`EngineeringRule`
and :class:`EngineeringTypesetter`.  :class:`EngineeringRule` is
different from :class:`ScientificRule` in that the *unit* string is
optional and handed over to :meth:`EngineeringRule.apply`, together with
the *exponent* which is optional as well.  :class:`EngineeringRule`
maintains independent columns for *numerical orders of magnitude*
(``10^xx``) and for a *unit* (e.g. ``mm``).

:class:`EngineeringTypesetter` can be asked by setting
*useprefixes=True* to emit *prefixes* (as k, M, G) for the *unit*
instead of showing powers of 10.  For orders of magnitude where no
prefix is defined, :meth:`EngineeringTypesetter.typeset_element` will
fall back to displaying a power of 10.

A comprehensive Test for the new code is provided, along with new Tests
of the scientific typesetter for *unit*, *padding* and for emitting '+'
signs in the nominal value and the exponent.