Preparation for using the :class:`Convention` Protocol
No longer importing :class:`{Scientific, Engineering,
Fixedpoint}Typesetter` in :mod:`typesetting.__init__`, instead importing
these from :mod:`upy2.__init__`.  With the old import scheme I ended up
with an import error which I didn't really understand, however I noticed
that it had to do with importing the Typesetters from within
:mod:`typesetting.__init__` when :mod:`core` requested just to import
:mod:`typesetting.protocol`.  With the changed code the importing works
fine.

Introduced with this commit is :class:`typesetting.protocol.Convention`,
which holds *printing conventions* about the *separator* and the
*negative sign*.  It is consulted by
:class:`typesetting.numbers.NumberTypesetter` to typeset negative
numbers, and by :class:`typesetting.scientific.ScientificTypesetter` in
:meth:`deduce_rule` about the *separator*.  Here, the *separator* can
still be overridden by specifying it upon initialisation of
:class:`ScientificTypesetter`, however, when *not* specifying it, it
does no longer default to ``' +- '``, but to the *convention* defined by
the :class:`Convention` session.

The *default convention* is to use ASCII ``-`` as negative sign and
``' +- '`` as the separator.

From the Typesetters, only :class:`ScientificTypesetter` respects the
Convention about the *separator* so far.