Improving coverage of the *unit* option and the *useprefix* option (engineering)
The *unit* option is now tested for all relative Typesetters
(scientific, fixedpoint and engineering); for the engineering typesetter
the behaviour with absence of *useprefixes* is covered as well.

The relative engineering typesetter's results with values too large or
too small for the prefixes are tested.

The behaviour of the relative engineering typesetter constructed with
only *useprefixes* (without *unit*) is tested as well.

I also observed some corner case when using large values for a relative
uncertainty of 0.1::

    >>> 1e23 / 1e24
    0.09999999999999999
    >>> 1e22 / 1e23
    0.1

leading to an relative uncertainty of ``1 +- 10.0 10^-2`` instead of ``1
+- 1.0 10^-1``.  This behaviour is reproduced and tested as an expected
failure in :class:`Test_TypesettingScientificRelativeU`.