First work on fixing the relative Typesetters
All of the relative Typesetters are flawed in that they use the
*precision* handed over on initialisation time directly for the nominal
value.  Instead, they should choose the precision of the nominal value
figure depending on the *uncertainty*.  This bug is crucial; the
reported nominal figures are misleading.

In the (absolute uncertainty) :class:`FixedpointTypesetter`, simplified
an ``elif`` clause to an ``else``, since its predicate is always True
when reached.

First work on fixing the RelativeFixedpointTypesetter, to make it
consider the precision *relative* to the uncertainty.  It now also
features, accordingly, an initialisation argument *infinite_precision*.

Reworked the Test of :class:`RelativeFixedpointTypesetter` together with
:class:`ScientificRelativeUTypesetter`.  Improved this Test slightly to
also cover the *padding* Convention and the new *inifinite_precision*
argument.

All this new code pathes are tested only lightly (interactively to see
the results); the Test Suite is not expected to pass yet.

To finish testing the new payload code, the results of using
:class:`RelativeFixedpointTypesetter` with
:class:`PixedpointRelativeUTypestter` needs to be reworked to reflect
the changes made.  After this, the remaining relative Typesetters
(scientific and engineering) need to be fixed.